+++ tests/test_apns.py
@@ -34,6 +34,8 @@ class ApnsTestCase(testutils.TestCase):
 
         # pretend our certificate exists
         patch("os.path.exists", lambda x: x == TEST_CERTFILE_PATH).start()
+        # Since no certificate exists, don't try to read it.
+        patch("sygnal.apnspushkin.ApnsPushkin._report_certificate_expiration").start()
         self.addCleanup(patch.stopall)
 
         super(ApnsTestCase, self).setUp()
@@ -55,17 +57,17 @@ class ApnsTestCase(testutils.TestCase):
         method.return_value = testutils.make_async_magic_mock(
             NotificationResult("notID", "200")
         )
-        self.sygnal.pushkins[PUSHKIN_ID].MAX_JSON_BODY_SIZE = 200
+        self.sygnal.pushkins[PUSHKIN_ID].MAX_JSON_BODY_SIZE = 240
 
         # Act
         self._request(self._make_dummy_notification([DEVICE_EXAMPLE]))
 
         # Assert
-        self.assertEquals(1, method.call_count)
+        self.assertEqual(1, method.call_count)
         ((notification_req,), _kwargs) = method.call_args
         payload = notification_req.message
 
-        self.assertLessEqual(len(apnstruncate.json_encode(payload)), 200)
+        self.assertLessEqual(len(apnstruncate.json_encode(payload)), 240)
 
     def test_payload_truncation_test_validity(self):
         """
@@ -84,7 +86,7 @@ class ApnsTestCase(testutils.TestCase):
         self._request(self._make_dummy_notification([DEVICE_EXAMPLE]))
 
         # Assert
-        self.assertEquals(1, method.call_count)
+        self.assertEqual(1, method.call_count)
         ((notification_req,), _kwargs) = method.call_args
         payload = notification_req.message
 
@@ -105,12 +107,13 @@ class ApnsTestCase(testutils.TestCase):
         resp = self._request(self._make_dummy_notification([DEVICE_EXAMPLE]))
 
         # Assert
-        self.assertEquals(1, method.call_count)
+        self.assertEqual(1, method.call_count)
         ((notification_req,), _kwargs) = method.call_args
 
-        self.assertEquals(
+        self.assertEqual(
             {
                 "room_id": "!slw48wfj34rtnrf:example.com",
+                "event_id": "$qTOWWTEL48yPm3uT-gdNhFcoHxfKbZuqRVnnWWSkGBs",
                 "aps": {
                     "alert": {
                         "loc-key": "MSG_FROM_USER_IN_ROOM_WITH_CONTENT",
@@ -121,13 +124,13 @@ class ApnsTestCase(testutils.TestCase):
                         ],
                     },
                     "badge": 3,
-                    "content-available": 1,
+                    "mutable-content": 1,
                 },
             },
             notification_req.message,
         )
 
-        self.assertEquals({"rejected": []}, resp)
+        self.assertEqual({"rejected": []}, resp)
 
     def test_rejection(self):
         """
@@ -144,8 +147,8 @@ class ApnsTestCase(testutils.TestCase):
         resp = self._request(self._make_dummy_notification([DEVICE_EXAMPLE]))
 
         # Assert
-        self.assertEquals(1, method.call_count)
-        self.assertEquals({"rejected": ["spqr"]}, resp)
+        self.assertEqual(1, method.call_count)
+        self.assertEqual({"rejected": ["spqr"]}, resp)
 
     def test_no_retry_on_4xx(self):
         """
@@ -162,8 +165,8 @@ class ApnsTestCase(testutils.TestCase):
         resp = self._request(self._make_dummy_notification([DEVICE_EXAMPLE]))
 
         # Assert
-        self.assertEquals(1, method.call_count)
-        self.assertEquals(502, resp)
+        self.assertEqual(1, method.call_count)
+        self.assertEqual(502, resp)
 
     def test_retry_on_5xx(self):
         """
@@ -181,4 +184,4 @@ class ApnsTestCase(testutils.TestCase):
 
         # Assert
         self.assertGreater(method.call_count, 1)
-        self.assertEquals(502, resp)
+        self.assertEqual(502, resp)